requirements.txt
annotated-types==0.7.0
anyio==4.9.0
blinker==1.9.0
firecrawl-py==2.13.0
openai==1.93.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.1.8
deprecation==2.1.0
Flask==3.1.1
flask-cors==6.0.0
Flask-SQLAlchemy==3.1.1
gotrue==2.12.2
greenlet==3.2.3
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
pluggy==1.6.0
postgrest==1.1.1
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.1
pytest-mock==3.14.1
python-dotenv==0.21.0
python-dateutil==2.9.0.post0
realtime==2.5.3
requests==2.32.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
storage3==0.12.0
StrEnum==0.4.15
supabase==2.16.0
supafunc==0.10.1
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.5.0
websockets==15.0.1
Werkzeug==3.1.3
